fix(IntegrationTests): fix /sign-out test + refactoring in tests
