+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '3.4.5'
+    id 'org.springframework.boot' version '3.5.5'
     id 'io.spring.dependency-management' version '1.1.7'
 }
 
@@ -29,7 +29,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.9'
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.10'
     implementation 'org.springframework.session:spring-session-data-redis'
     implementation 'io.minio:minio:8.5.17'
     implementation "org.flywaydb:flyway-core:11.8.0"
@@ -50,9 +50,6 @@ dependencies {
     testImplementation("org.testcontainers:postgresql:1.21.3")
     testImplementation("com.redis.testcontainers:testcontainers-redis:1.6.4")
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
-
-//    Copy
 
     testImplementation 'org.junit.platform:junit-platform-engine:1.13.0-M2'
     testImplementation 'org.junit.platform:junit-platform-commons:1.13.0-M2'
