+++ src/main/java/org/fizz_buzz/cloud/config/SecurityConfig.java
@@ -31,7 +31,6 @@ public class SecurityConfig {
 
         http.authorizeHttpRequests(authorize -> authorize
                         .requestMatchers("/api/v1/auth/sign-up", "/api/v1/auth/sign-in").permitAll()
-                        .requestMatchers("/api/v1/auth/sign-out").fullyAuthenticated()
                         .anyRequest().authenticated())
                 // Включаем CORS и разрешаем все источники, методы и заголовки
                 .cors(cors -> cors.configurationSource(corsConfigurationSource()))
