--- src/test/java/org/fizz_buzz/cloud/CloudApplicationTests.java
@@ -1,13 +1,0 @@
-package org.fizz_buzz.cloud;
-
-import org.junit.jupiter.api.Test;
-//import org.springframework.boot.test.context.SpringBootTest;
-
-//@SpringBootTest
-class CloudApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
