--- src/test/java/org/fizz_buzz/cloud/controller/AuthControllerTest.java
@@ -1,30 +1,0 @@
-package org.fizz_buzz.cloud.controller;
-
-import org.fizz_buzz.cloud.repository.UserRepository;
-import org.hibernate.annotations.Array;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.boot.test.context.SpringBootTest;
-//import org.springframework.test.web.servlet.MockMvc;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-//@SpringBootTest
-class AuthControllerTest {
-
-//    @Autowired
-//    private UserRepository userRepository;
-//
-//    private MockMvc mockMvc;
-
-    @AfterEach
-    void tearDown() {
-
-    }
-
-    @Test
-    void signUp() {
-//        mockMvc.
-    }
-}
